refactor: update precompile dispatch
