+++ crates/precompiles/Cargo.toml
@@ -37,6 +37,7 @@ sha2 = { version = "0.10.9", default-features = false }
 rand = { version = "0.8.5", default-features = false }
 
 # misc
+paste = "1.0.15"
 tracing = { version = "0.1.41", default-features = false }
 
 [dev-dependencies]
