+++ crates/precompiles/src/chaum_pedersen/mod.rs
@@ -10,9 +10,10 @@
 
 use alloc::{borrow::Cow, vec::Vec};
 
-use alloy_evm::precompiles::{DynPrecompile, Precompile, PrecompileInput};
-use alloy_primitives::{Address, Bytes, address};
-use alloy_sol_types::SolCall;
+mod dispatch;
+
+use alloy_evm::precompiles::{DynPrecompile, Precompile};
+use alloy_primitives::{Address, address};
 use k256::{
     AffinePoint, ProjectivePoint, Scalar,
     elliptic_curve::{
@@ -20,8 +21,7 @@ use k256::{
         sec1::{FromEncodedPoint, ToEncodedPoint},
     },
 };
-use revm::precompile::{PrecompileId, PrecompileOutput, PrecompileResult};
-use tracing::{debug, warn};
+use revm::precompile::PrecompileId;
 
 /// Chaum-Pedersen Verify precompile address on Zone L2.
 pub const CHAUM_PEDERSEN_VERIFY_ADDRESS: Address =
@@ -40,17 +40,21 @@ alloy_sol_types::sol! {
         bytes32 c;
     }
 
-    /// Verify a Chaum-Pedersen proof of correct ECDH shared secret derivation.
-    function verifyProof(
-        bytes32 ephemeralPubX,
-        uint8 ephemeralPubYParity,
-        bytes32 sharedSecret,
-        uint8 sharedSecretYParity,
-        bytes32 sequencerPubX,
-        uint8 sequencerPubYParity,
-        ChaumPedersenProof proof
-    ) external view returns (bool valid);
+    interface IChaumPedersenVerify {
+        /// Verify a Chaum-Pedersen proof of correct ECDH shared secret derivation.
+        function verifyProof(
+            bytes32 ephemeralPubX,
+            uint8 ephemeralPubYParity,
+            bytes32 sharedSecret,
+            uint8 sharedSecretYParity,
+            bytes32 sequencerPubX,
+            uint8 sequencerPubYParity,
+            ChaumPedersenProof proof
+        ) external view returns (bool valid);
+    }
 }
+
+pub use IChaumPedersenVerify::verifyProofCall;
 
 /// Chaum-Pedersen DLOG equality proof verification precompile.
 ///
@@ -64,50 +68,6 @@ alloy_sol_types::sol! {
 /// - `c' = keccak256(G, ephemeralPub, pubSeq, sharedSecretPoint, R1, R2)`
 /// - Check: `c == c'`
 pub struct ChaumPedersenVerify;
-
-impl Precompile for ChaumPedersenVerify {
-    fn precompile_id(&self) -> &PrecompileId {
-        &CP_PRECOMPILE_ID
-    }
-
-    fn call(&self, input: PrecompileInput<'_>) -> PrecompileResult {
-        let data = input.data;
-        if data.len() < 4 {
-            return Ok(PrecompileOutput::revert(0, Bytes::new(), input.reservoir));
-        }
-
-        let selector: [u8; 4] = data[..4].try_into().expect("len >= 4");
-        if selector != verifyProofCall::SELECTOR {
-            warn!(target: "zone::precompile", ?selector, "ChaumPedersenVerify: unknown selector");
-            return Ok(PrecompileOutput::revert(0, Bytes::new(), input.reservoir));
-        }
-
-        debug!(target: "zone::precompile", "ChaumPedersenVerify: verifyProof");
-
-        let call = match verifyProofCall::abi_decode(data) {
-            Ok(call) => call,
-            Err(_) => return Ok(PrecompileOutput::revert(0, Bytes::new(), input.reservoir)),
-        };
-
-        let valid = verify_chaum_pedersen(
-            &call.ephemeralPubX.0,
-            call.ephemeralPubYParity,
-            &call.sharedSecret.0,
-            call.sharedSecretYParity,
-            &call.sequencerPubX.0,
-            call.sequencerPubYParity,
-            &call.proof.s.0,
-            &call.proof.c.0,
-        );
-
-        let encoded = verifyProofCall::abi_encode_returns(&valid);
-        Ok(PrecompileOutput::new(
-            CP_VERIFY_GAS,
-            encoded.into(),
-            input.reservoir,
-        ))
-    }
-}
 
 impl ChaumPedersenVerify {
     /// Convert into a [`DynPrecompile`] for registration in a [`PrecompilesMap`].
